approximate solution to spindle-cusp field
I tinkered around with a way to solve for the pressure equilibrium.
However, there are two problems with this basic approach. 1) I don't
have the correct shape. I am using a simple circular arc for the shape,
and assuming certain aspect ratios. So no manner of currents will give
the true solution. 2) I am guessing blindly at current densities for the
boundary. I can tinker around all day and it won't be right. To solution
needs to be nearly exact to be  close enough.

There are two possibilities. 1) do like I did before, and create an
iterative solver for the boundary shape, and current density. Moving the
boundary in and out until it finds equilibrium for constant pressure. 2)
Try to use the plasma simulation itself to get the solution by
connecting plasma back to field through a dynamic field solver.

For now, option 1 seems like the most pragmatic choice to go. I am
fairly certain it can success, while option 2 would be more involved to
implement and may not give a good answer.